Move CoreComponents to components file
