+++ src/components/inflationStats.ts
@@ -1,0 +1,5 @@
+import { BaseComponent } from "./component";
+
+export interface InflationStats extends BaseComponent<"inflationStats"> {
+  basketPrices: number[];
+}
